tests/pyproject.toml
[project]
name = "ai-voice-e2e-tests"
version = "0.1.0"
description = "System-wide end-to-end tests for AI Voice Tool"
authors = [
    { name = "Niklas", email = "niklas@alyst.se" }
]
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.40.71",
    "httpx>=0.28.1",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["e2e"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "e2e: End-to-end tests that require deployed infrastructure",
]
